requirements.txt
channels==4.0.0
Django==4.1.10
django-countries==7.5.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-phonenumber-field==7.0.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
phonenumbers==8.13.7
pillow==10.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0
whitenoise==6.4.0
